[builtin] envコマンド実装

+++ tests/grademe.sh
@@ -41,6 +41,7 @@ run_all_tests () {
 		run_tests "shlvl"
 		run_tests "pwd"
 		run_tests "cd"
+		run_tests "env"
 	fi
 }
 
@@ -66,11 +67,17 @@ print_usage () {
 
 cleanup () {
 	rm -f ${BASH_STDOUT_FILE} ${BASH_STDERR_FILE} ${MINISHELL_STDOUT_FILE} ${MINISHELL_STDERR_FILE}
-	rm -fr ${TEST_DIR}
+	if [ -e "${TEST_DIR}" ]; then
+		chmod -R 777 ${TEST_DIR}
+		rm -fr ${TEST_DIR}
+	fi
 }
 
 prepare_test_dir () {
-	rm -fr ${TEST_DIR}
+	if [ -e "${TEST_DIR}" ]; then
+		chmod -R 777 ${TEST_DIR}
+		rm -fr ${TEST_DIR}
+	fi
 	mkdir -p ${TEST_DIR}
 	cd ${TEST_DIR}
 }
